Refactor Map modules and classes
d_rats/map/__init__.py:
  Add MapBottomPanel, MapMarkerList, MapStatusBox, and MapZoomControls.

d_rats/map/mapbottompanel.py:
  Classes for the Map Bottom Panel.

d_rats/map/mapmarkerlist.py:
  Class for Map Marker List.

d_rats/map/mapstatusbox.py:
  Class for Map Status Box.

d_rats/map/mapwidget.py;
  Minor fixes.

d_rats/map/mapwindow.py:
  Remove code for constructing MapBottomPanel, MapMarkerList,
  MapStatusBox, and MapZoomControls

d_rats/map/mapzoomcontrols.py:
  Class for Map Zoom Controls
