+++ d_rats/map/__init__.py
@@ -1,4 +1,8 @@
 '''Package to handle Maps for D-Rats.'''
 
-from .mapwindow import MapWindow as Window
+from .mapbottompanel import MapBottomPanel as BottomPanel
+from .mapmarkerlist import MapMarkerList as MarkerList
+from .mapstatusbox import MapStatusBox as StatusBox
 from .mapwidget import MapWidget as Widget
+from .mapwindow import MapWindow as Window
+from .mapzoomcontrols import MapZoomControls as ZoomControls
